% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentCSV2DF.R
\name{alignmentCSV2DF}
\alias{alignmentCSV2DF}
\title{Convert an Alignment CSV from BLAST to a Dataframe}
\usage{
alignmentCSV2DF(alignmentFile)
}
\arguments{
\item{alignmentFile}{The file location of the CSV you want to convert to a compatible dataframe.}
}
\description{
This function allows you to convert a CSV file generated by a BLAST search to a dataframe compatible with the joinSpacerDFandAlignmentDF function. The function adds the appropriate headers that are not preserved by BLAST export.
}
\examples{
alignmentCSV2DF()
}
