% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2fasta.R
\name{df2fasta}
\alias{df2fasta}
\title{Convert a Spacer Dataframe to a FASTA file}
\usage{
df2fasta(spacerDataFrame)
}
\arguments{
\item{spacerDataFrame}{The appropriately formatted data frame containing the CRISPR array spacers and associated data.}
}
\description{
This function allows you to convert a dataframe containing the spacer data to a FASTA file with annotation nomenclature compatible with the rest of this package. The FASTA is written to the current file location with file name "_spacers.fasta" where the blank is filled with the systemName as defined by the setCRISPRInfo function.
}
\examples{
df2fasta()
}
