% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTA2AlignmentDF.R
\name{FASTA2AlignmentDF}
\alias{FASTA2AlignmentDF}
\title{Submit spacer FASTA to BLAST and format the results}
\usage{
FASTA2AlignmentDF(FASTAfile)
}
\arguments{
\item{FASTAfile}{The file location of the FASTA you want to submit to BLAST.}
}
\description{
This function allows you to programatically submit your spacer FASTA to BLAST and retrieve the results. The results are then parsed and formatted for use with the joinSpacerDFandAlignmentDF function.
}
\examples{
FASTA2AlignmentDF()
}
