% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinSpacerDFandAlignmentDF.R
\name{joinSpacerDFandAlignmentDF}
\alias{joinSpacerDFandAlignmentDF}
\title{Join Spacer and Alignment Dataframes}
\usage{
joinSpacerDFandAlignmentDF(alignmentDF, spacerDF, accessionDatabaseLocation)
}
\arguments{
\item{alignmentDF}{The dataframe containing the CRISPR array spacer information.}

\item{spacerDF}{The dataframe containing the spacer alignment information.}

\item{accessionDatabaseLocation}{The location of your Taxonomizer SQL database.}
}
\description{
This function allows you to join a dataframe containing the CRISPR array spacer information with a dataframe containing the results of the BLAST alignment. Produces a dataframe with the same number of rows as the alignment dataframe and the spacer information added as additional columns. This function also takes the accession numbers for the alignment and adds a column with the Taxa IDs and organism name. Taxa IDs
}
\examples{
joinSpacerDFandAlignmentDF(alignmentDF = yourAlignmentDataframe, spacerDF = yourSpacerDataframe, accessionDatabaseLocation = "theLocationOfYourTaxonomizerDatabase")
}
