% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCRISPRInfo.R
\name{setCRISPRInfo}
\alias{setCRISPRInfo}
\title{Set the Name of the CRISPR System}
\usage{
setCRISPRInfo(genus, species, strain, crisprSystemNumber)
}
\arguments{
\item{genus}{A string value that indicates the genus name of the organism encoding the CRISPR array. Make sure to capitalize the first letter of the genus name.}

\item{species}{A string value that indicates the species name of the organism encoding the CRISPR array. Make sure not to capitalize the first letter of the species name.}

\item{strain}{A string value that indicates the strain name of the organism encoding the CRISPR array.}

\item{crisprSystemNumber}{A string value that will act as an identifier to distinguish multiple CRISPR systems in the same organism.}
}
\description{
This function allows you to set the genus, species, and strain of the organism that encodes the CRISPR system. It also sets a unique identifier for the CRISPR system as there can be multiple per organism. These values will be used by other functions to name files.
}
\examples{
Given a CRISPR system in Bacillus halodurans C-125 that you designate as "1", you would enter:
setCRISPRInfo(genus = "Bacillus", species = "halodurans", strain = "C-125", crisprSystemNumber = "1")
}
