% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit2Phaster.R
\name{submit2Phaster}
\alias{submit2Phaster}
\title{Submits Alignment Hits for Prophage Prediction}
\usage{
submit2Phaster(joinedData = joineddata, startPosition = 1)
}
\arguments{
\item{joinedData}{A dataframe containing the joined spacer and alignment information. Defaults to joineddata, the output of the joinSpacerDFandAlignmentDF function.}

\item{startPosition}{A value to indicate where in the list of accession numbers to start. Defaults to 1.}
}
\description{
This function allows you to submit the accession numbers of organisms identified by BLAST alignment to Phaster for prophage prediction. This process can take a bit of time to run and depends on the number of sequences being submitted.
}
\examples{
submit2Phaster()
}
